+++ src/main/java/org/folio/linked/data/integration/kafka/listener/handler/ExternalEventHandler.java
@@ -1,0 +1,6 @@
+package org.folio.linked.data.integration.kafka.listener.handler;
+
+public interface ExternalEventHandler<T> {
+
+  void handle(T event);
+}