+++ src/main/java/org/folio/linked/data/service/resource/events/ResourceEventsPublisher.java
@@ -4,11 +4,8 @@ import org.folio.linked.data.model.entity.Resource;
 import org.folio.linked.data.service.resource.graph.SaveGraphResult;
 
 public interface ResourceEventsPublisher {
-  void emitEventsForCreate(SaveGraphResult saveGraphResult);
 
-  void emitEventsForUpdate(SaveGraphResult saveGraphResult);
-
-  void emitEventsForUpdate(Resource oldResource, SaveGraphResult saveGraphResult);
+  void emitEventsForCreateAndUpdate(SaveGraphResult saveGraphResult, Resource oldResource);
 
   void emitEventForDelete(Resource deletedResource);
 }